Create ColumnInput
